Template to get going with LL
Enter a linked list of unknown size where every element is an int and each node is separated by either '>' or '|'. The linked list should end with a null. Used a buffer to get the input from stdin using fgets. This linked list is broken into tokens using strtok and nodes are created dynamically by passing each of the integer tokens from the user string in the buffer.