Added: Merge two sorted lists - recursive
say 
1 2
3 4 
are the given sorted lists.
the smallest element will be one of the head nodes' data. 
say head1.
thus, keep it aside and update it's next by assigning it to the next smaller element in both the lists. this element would also be at either of the head nodes. so, head1->next = (head1->next, head2);
proceed further in the same way.