Implement disjoint set ADT using BST
We need to keep track of the disjoint sets that constitute to an equivalence class. BST is used because it is as efficient as it can be except for the case of a skewed tree. BST is formed with all the elements initially, considering each element to be an equiv. class.